requirements.txt
gradio
requests
python-dotenv
wikipedia
python-dotenv
langchain-tavily
langchain
langchain-community
arxiv
langchain-google-genai
langgraph
gradio[oauth]
pymupdf